Update README.md
## Challenge Description
The goal is to design a dashboard that displays users based on their social reputation within a community.The dashboard should show the most active and influential members and allow users to follow them. The reputation rankings should be transparent, and the user interface should clearly explain the criteria and weighting used to generate them.

## Proposed solution:
Our proposed solution is a Web 3 app dashboard. We designed a dashboard which spools data via chains/pipelines to generate analytics and insights using Subsquid. Our solution focuses on building a design which can display users based on their reputation rankings. 

The project deliverables.
1. The web 3 app dashboard. 
2. A landing page for the product.
3. Documentation of the product. 

## About product
GPX is a web 3 social network platform that spools data via pipelines generating analytics and insights on the activities of users in the community. 

Find project presentation [here]